requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.25.2
python-dotenv==1.0.0

# Google Cloud
google-cloud-speech==2.21.0
google-cloud-translate==3.12.1
google-generativeai==0.3.2

# Audio processing
soundfile==0.12.1
pydub==0.25.1
numpy==1.24.4

# Utilities
requests==2.31.0
aiofiles==23.2.0
aiohttp==3.9.1

# AI/ML
openai==1.3.7
